+++ src/index.ts
@@ -38,3 +38,4 @@ export * from './components/typography';
 export * from './components/notification';
 export * from './components/link';
 export * from './components/root-element';
+export * from './components/spinner';